requirements.txt
streamlit>=1.20.0
groq>=0.6.0
python-dotenv>=1.0.0
